don't apply accessor generator role multiple times
Since we are applying this to an object, Moo/Role::Tiny lets us apply it
multiple times.  This leads to issues with its composed classes used as
an optimization.  This seems like a bug in Role::Tiny, but we can avoid
easily by just not applying the role more than once.